Decouple MacSettingsModel from AppMenu
